LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securityviolationpolicy-block-frame-using-child-src-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-fail.html because it does not appear in the child-src directive of the Content Security Policy.
Check that a SecurityPolicyViolationEvent is fired upon blocking an frame by the child-src directive.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Kicking off the tests:
PASS window.e.documentURI is "http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/securityviolationpolicy-block-frame-using-child-src.html"
PASS window.e.referrer is ""
PASS window.e.blockedURI is "http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-fail.html"
PASS window.e.violatedDirective is "child-src 'none'"
PASS window.e.effectiveDirective is "child-src"
PASS window.e.originalPolicy is "child-src 'none'"
PASS window.e.sourceFile is "http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/securityviolationpolicy-block-frame-using-child-src.html"
PASS window.e.lineNumber is 27
PASS window.e.columnNumber is 30
PASS window.e.statusCode is 200
PASS successfullyParsed is true

TEST COMPLETE
